hatchling does DOT

+++ client/include/world/crop/Crop.h
@@ -12,7 +12,7 @@
 class Crop : public Entity {
 protected:
     float time_since_planted = 0.f;
-    int health;
+    float health;
     mutable sf::Sprite sprite;
 public:
     explicit Crop(Space& space, sf::Vector2f location);
@@ -25,7 +25,7 @@ public:
 	
 	virtual float get_time_to_maturity() const = 0;
 
-    void damage(int value);
+    void damage(float value);
 
     bool shouldBeRemoved() const override;
 };
